OpenAI-DotNet 7.3.4
- Fixed AudioTranslationRequest.Temperature type. int? -&gt; float?

+++ OpenAI-DotNet/Audio/AudioTranslationRequest.cs
@@ -33,7 +33,7 @@ namespace OpenAI.Audio
             string model = null,
             string prompt = null,
             AudioResponseFormat responseFormat = AudioResponseFormat.Json,
-            int? temperature = null)
+            float? temperature = null)
             : this(File.OpenRead(audioPath), Path.GetFileName(audioPath), model, prompt, responseFormat, temperature)
         {
         }
@@ -70,7 +70,7 @@ namespace OpenAI.Audio
             string model = null,
             string prompt = null,
             AudioResponseFormat responseFormat = AudioResponseFormat.Json,
-            int? temperature = null)
+            float? temperature = null)
         {
             Audio = audio;
 
@@ -121,7 +121,7 @@ namespace OpenAI.Audio
         /// the model will use log probability to automatically increase the temperature until certain thresholds are hit.<br/>
         /// Defaults to 0
         /// </summary>
-        public int? Temperature { get; }
+        public float? Temperature { get; }
 
         private void Dispose(bool disposing)
         {
